Relocating the homebridge services.
