+++ src/services/batteryService.ts
@@ -1,7 +1,7 @@
 import { Characteristic, Service } from 'homebridge';
-import { InvalidStateError } from '../../errors/invalidStateError';
+import { InvalidStateError } from '../errors/invalidStateError';
 
-import { Activity, Battery, MowerState } from '../../model';
+import { Activity, Battery, MowerState } from '../model';
 import { AbstractAccessoryService } from './abstractAccessoryService';
 
 /**
