+++ test/services/policies/scheduleEnabledPolicy.spec.ts
@@ -1,6 +1,6 @@
-import { InvalidStateError } from '../../../../src/errors/invalidStateError';
-import { Activity, Calendar, Mode, RestrictedReason, State } from '../../../../src/model';
-import { DeterministicScheduleEnabledPolicy } from '../../../../src/services/homebridge/policies/scheduleEnabledPolicy';
+import { InvalidStateError } from '../../../src/errors/invalidStateError';
+import { Activity, Calendar, Mode, RestrictedReason, State } from '../../../src/model';
+import { DeterministicScheduleEnabledPolicy } from '../../../src/services/policies/scheduleEnabledPolicy';
 
 describe('DeterministicScheduleEnabledPolicy', () => {
     let target: DeterministicScheduleEnabledPolicy;
